+++ Changelog.md
@@ -1,3 +1,13 @@
+### v 0.1.8
+
+#### Enable and Disable Archiving for Accounts
+
+```
+ post /accounts/:id/archive/disable
+ post /accounts/:id/archive/enable, { cos_id: cos_id, archive_name: archive_name}
+```
+
+
 ### v 0.1.4
 
 #### new endpoint `/distribution_lists/:id/add_members`